fix(SR-9): Apply csrf checks to supporting document routes
