+++ gef-ui/server/routes/__tests__/supporting-information-upload.test.js
@@ -1,5 +1,3 @@
-import multer from 'multer';
-
 import { validateToken, validateBank } from '../../middleware';
 
 jest.mock('multer', () => () => ({
@@ -7,12 +5,10 @@ jest.mock('multer', () => () => ({
   single: jest.fn(),
 }));
 
-const getSpy = jest.fn();
 const postSpy = jest.fn();
 jest.doMock('express', () => ({
   Router() {
     return {
-      get: getSpy,
       post: postSpy,
     };
   },
@@ -21,7 +17,7 @@ jest.doMock('express', () => ({
 describe('Routes', () => {
   beforeEach(() => {
     // eslint-disable-next-line global-require
-    require('../supporting-information');
+    require('../supporting-information-upload');
   });
 
   afterEach(() => {
@@ -29,14 +25,14 @@ describe('Routes', () => {
   });
 
   it('Sets up all routes', () => {
-    expect(getSpy).toHaveBeenCalledWith(
-      '/application-details/:dealId/supporting-information/document/:documentType',
-      [validateToken, validateBank, expect.any(Function)],
+    expect(postSpy).toHaveBeenCalledWith(
+      '/application-details/:dealId/supporting-information/document/:documentType/upload',
+      [expect.any(Function), validateToken, validateBank, expect.any(Function), expect.any(Function)],
       expect.any(Function),
     );
     expect(postSpy).toHaveBeenCalledWith(
-      '/application-details/:dealId/supporting-information/document/:documentType',
-      [validateToken, validateBank, multer().array('documents', 2)],
+      '/application-details/:dealId/supporting-information/document/:documentType/delete',
+      [validateToken, validateBank, expect.any(Function)],
       expect.any(Function),
     );
   });
