+++ gef-ui/server/routes/index.js
@@ -3,6 +3,7 @@ const sessionUserToLocals = require('../middleware/sessionUserToLocals');
 const mandatoryCriteriaRoutes = require('./mandatory-criteria');
 const nameApplicationRoutes = require('./name-application');
 const applicationDetailsRoutes = require('./application-details');
+const supportingInformationRoutes = require('./supporting-information');
 const applicationSubmissionRoutes = require('./application-submission');
 const applicationAbandonRoutes = require('./application-abandon');
 const submitToUkefRoutes = require('./submit-to-ukef');
@@ -38,6 +39,7 @@ router.use(mandatoryCriteriaRoutes);
 router.use(nameApplicationRoutes);
 router.use(eligibleAutomaticCoverRoutes);
 router.use(applicationDetailsRoutes);
+router.use(supportingInformationRoutes);
 router.use(applicationSubmissionRoutes);
 router.use(applicationAbandonRoutes);
 router.use(submitToUkefRoutes);
