--- gef-ui/server/routes/supporting-documents.route.js
@@ -1,7 +1,0 @@
-const express = require('express');
-const supportingInformation = require('./supporting-information');
-
-const router = express.Router();
-router.use(supportingInformation);
-
-module.exports = router;
